chore(demo-app): move services into dedicated folder
