+++ apps/chill-viking-ng-libs/src/app/services.ts
@@ -1,0 +1,2 @@
+export * from './services/page-meta-data.service';
+export * from './services/window-router.service';
