+++ apps/chill-viking-ng-libs/src/main.ts
@@ -1,8 +1,8 @@
 import { bootstrapApplication } from '@angular/platform-browser';
 import { provideRouter, TitleStrategy } from '@angular/router';
-import { PageMetaDataService } from './app/page-meta-data.service';
 import { RootComponent } from './app/root.component';
 import { rootRouterFeatures, rootRoutes } from './app/routes';
+import { PageMetaDataService } from './app/services';
 
 bootstrapApplication(RootComponent, {
   providers: [
